feat: auth 모듈 MySQL → PostgreSQL 마이그레이션 및 prod 프로파일 추가

+++ auth/src/main/java/dev/montyoh/auth/domain/repository/RoleRepository.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 public interface RoleRepository {
 
-    List<Role> findAllByIdIn(List<Integer> ids);
+    List<Role> findAllByIdIn(List<Long> ids);
 
     Role save(Role role);
 
     List<Role> saveAll(List<Role> roles);
 
-    Role findById(int id);
+    Role findById(Long id);
 }
